+++ packages/core/src/index.ts
@@ -75,7 +75,7 @@ export function css<VariantsMap extends QuarkVariantsMap>(
       }
     }
 
-    return classNames.join(' ')
+    return cleanMultiLine(classNames.join(' '))
   }
 
   return Object.assign(css, {
@@ -99,5 +99,7 @@ const normalize = (key: string | boolean | null | undefined): string => {
   //If falsey, return 'null' as the prop key, if true, return 'true'
   return !key ? 'null' : key.toString()
 }
+
+const cleanMultiLine = (str: string) => str.replace(/\s+/g, ' ').trim()
 
 const compoundPropKeywords = new Set(['value', 'negate'])